# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Training Center",
        "description": "Educational Lab",
        "license": {
            "name": "Apache 2.4",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo TC API Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "489c38c55debbddbfae2b1d58cab5257",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register",
                "operationId": "d03328a9330fe717d677f5fce93140d6",
                "requestBody": {
                    "description": "Pass user registration credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "firstName",
                                    "lastName",
                                    "phone",
                                    "roleId"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "PassWord12345"
                                    },
                                    "firstName": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Smith"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "094415263"
                                    },
                                    "taxIdentityNumber": {
                                        "type": "integer",
                                        "example": 5720125
                                    },
                                    "roleId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}